Bug Fixes - added distribute_workers so that goes back to normal - set first tumor target to 11. 12 worked on starting base at bottom right, but failed at top left (I think)
