+++ combatmgr.py
@@ -20,3 +20,27 @@ class CombatManager:
 
     async def manage(self):
         pass
+
+
+
+    def get_pos_around_unit(
+        self, unit, min_range=0, max_range=500, step_size=1, loc_amt=32
+    ):
+        """
+        # e.g. loc_amt=4 would only consider 4 points: north, west, east, south
+        """
+        loc = unit.position.to2
+        # loc = unit
+        positions = [
+            Point2(
+                (
+                    loc.x + distance * math.cos(math.pi * 2 * alpha / loc_amt),
+                    loc.y + distance * math.sin(math.pi * 2 * alpha / loc_amt),
+                )
+            )
+            for alpha in range(
+                loc_amt
+            )  # alpha is the angle here, locationAmount is the variable on how accurate the attempts look like a circle (= how many points on a circle)
+            for distance in range(min_range, max_range + 1)
+        ]  # distance depending on minrange and maxrange
+        return positions
